requirements.txt
python-dotenv
streamlit
chromadb
langchain
langchain-community
langchain-chroma
langchain-ollama
langchain-google-genai
pymysql
beautifulsoup4
lxml
tqdm
streamlit-option-menu
